requirements.txt
PyYAML
pyzmq
colorlog
pygame
numpy
networkx
shapely
